# backend/internal/handlers/stub_handlers.go
package handlers

import (
	"time"

	"github.com/defi-dashboard/backend/pkg/errors"
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
)

// Yield Handlers

// GetYieldPools handles GET /yield/pools
func GetYieldPools(c *fiber.Ctx) error {
	// TODO: Implement actual yield pool fetching
	// Mock response for now
	
	pools := []fiber.Map{
		{
			"id":       "uniswap-v3-weth-usdc",
			"protocol": "Uniswap V3",
			"name":     "WETH/USDC 0.05%",
			"chainId":  1,
			"apy":      12.5,
			"tvl":      125000000,
			"tokens": []fiber.Map{
				{
					"address":  "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
					"symbol":   "WETH",
					"name":     "Wrapped Ether",
					"decimals": 18,
				},
				{
					"address":  "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
					"symbol":   "USDC",
					"name":     "USD Coin",
					"decimals": 6,
				},
			},
			"risks": []string{"impermanent_loss", "smart_contract"},
		},
		{
			"id":       "aave-v3-usdc",
			"protocol": "Aave V3",
			"name":     "USDC Supply",
			"chainId":  1,
			"apy":      4.2,
			"tvl":      450000000,
			"tokens": []fiber.Map{
				{
					"address":  "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
					"symbol":   "USDC",
					"name":     "USD Coin",
					"decimals": 6,
				},
			},
			"risks": []string{"smart_contract"},
		},
	}

	return c.JSON(fiber.Map{
		"data": pools,
		"meta": fiber.Map{
			"page":       1,
			"limit":      20,
			"total":      2,
			"totalPages": 1,
		},
	})
}

// GetYieldPositions handles GET /yield/positions/:address
func GetYieldPositions(c *fiber.Ctx) error {
	// TODO: Implement actual yield position fetching
	// Mock response for now
	
	return c.JSON(fiber.Map{
		"totalValue": 15234.56,
		"totalPnl":   1234.56,
		"positions": []fiber.Map{
			{
				"pool": fiber.Map{
					"id":       "uniswap-v3-weth-usdc",
					"protocol": "Uniswap V3",
					"name":     "WETH/USDC 0.05%",
					"chainId":  1,
					"apy":      12.5,
				},
				"balance":        "1000000000000000000",
				"balanceUsd":     10234.56,
				"rewards":        []fiber.Map{},
				"entryTime":      time.Now().Add(-30 * 24 * time.Hour),
				"pnl":            234.56,
				"pnlPercentage":  2.35,
			},
		},
	})
}

// Bridge Handlers

// GetBridgeRoutes handles POST /bridge/routes
func GetBridgeRoutes(c *fiber.Ctx) error {
	// TODO: Implement actual bridge route calculation
	// Mock response for now
	
	var req fiber.Map
	if err := c.BodyParser(&req); err != nil {
		return errors.BadRequest("Invalid request body")
	}

	return c.JSON(fiber.Map{
		"routes": []fiber.Map{
			{
				"id":       uuid.New().String(),
				"fromChain": req["fromChain"],
				"toChain":   req["toChain"],
				"fromToken": fiber.Map{
					"address":  req["fromToken"],
					"symbol":   "USDC",
					"decimals": 6,
				},
				"toToken": fiber.Map{
					"address":  "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
					"symbol":   "USDC",
					"decimals": 6,
				},
				"bridge":        "Stargate",
				"estimatedTime": 300, // 5 minutes
				"fee":           2.5,
				"feeUsd":        2.5,
			},
		},
	})
}

// Analytics Handlers

// ExportPnL handles GET /analytics/pnl/:address
func ExportPnL(c *fiber.Ctx) error {
	// TODO: Implement actual P&L calculation and export
	// Mock response for now
	
	format := c.Query("format", "json")
	
	if format == "csv" {
		c.Set("Content-Type", "text/csv")
		c.Set("Content-Disposition", "attachment; filename=pnl_export.csv")
		return c.SendString("Date,Type,Asset,Amount,Price,PnL,Fees\n2024-01-01,swap,WETH,1.5,2300,150.50,5.00")
	}

	return c.JSON(fiber.Map{
		"startDate":          c.Query("startDate"),
		"endDate":            c.Query("endDate"),
		"totalPnl":           1234.56,
		"totalPnlPercentage": 12.34,
		"transactions": []fiber.Map{
			{
				"date":   time.Now().Add(-24 * time.Hour),
				"type":   "swap",
				"asset":  "WETH",
				"amount": 1.5,
				"price":  2300,
				"pnl":    150.50,
				"fees":   5.00,
			},
		},
	})
}

// Alert Handlers

// GetAlerts handles GET /alerts
func GetAlerts(c *fiber.Ctx) error {
	// TODO: Implement actual alert fetching
	return c.JSON(fiber.Map{
		"data": []fiber.Map{
			{
				"id":     uuid.New().String(),
				"type":   "price_above",
				"status": "active",
				"target": fiber.Map{
					"type":       "token",
					"identifier": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
					"chainId":    1,
				},
				"conditions": fiber.Map{
					"price": 3000,
				},
				"notification": fiber.Map{
					"email": true,
				},
				"createdAt": time.Now().Add(-7 * 24 * time.Hour),
			},
		},
		"meta": fiber.Map{
			"page":       1,
			"limit":      20,
			"total":      1,
			"totalPages": 1,
		},
	})
}

// CreateAlert handles POST /alerts
func CreateAlert(c *fiber.Ctx) error {
	// TODO: Implement actual alert creation
	var req fiber.Map
	if err := c.BodyParser(&req); err != nil {
		return errors.BadRequest("Invalid request body")
	}

	return c.Status(201).JSON(fiber.Map{
		"id":           uuid.New().String(),
		"type":         req["type"],
		"status":       "active",
		"target":       req["target"],
		"conditions":   req["conditions"],
		"notification": req["notification"],
		"createdAt":    time.Now(),
	})
}

// GetAlert handles GET /alerts/:alertId
func GetAlert(c *fiber.Ctx) error {
	alertId := c.Params("alertId")
	// TODO: Implement actual alert fetching
	return c.JSON(fiber.Map{
		"id":     alertId,
		"type":   "price_above",
		"status": "active",
		"target": fiber.Map{
			"type":       "token",
			"identifier": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
			"chainId":    1,
		},
		"conditions": fiber.Map{
			"price": 3000,
		},
		"notification": fiber.Map{
			"email": true,
		},
		"createdAt": time.Now().Add(-7 * 24 * time.Hour),
	})
}

// UpdateAlert handles PUT /alerts/:alertId
func UpdateAlert(c *fiber.Ctx) error {
	alertId := c.Params("alertId")
	// TODO: Implement actual alert update
	var req fiber.Map
	if err := c.BodyParser(&req); err != nil {
		return errors.BadRequest("Invalid request body")
	}

	return c.JSON(fiber.Map{
		"id":           alertId,
		"type":         "price_above",
		"status":       req["status"],
		"conditions":   req["conditions"],
		"notification": req["notification"],
		"updatedAt":    time.Now(),
	})
}

// DeleteAlert handles DELETE /alerts/:alertId
func DeleteAlert(c *fiber.Ctx) error {
	// TODO: Implement actual alert deletion
	return c.SendStatus(204)
}

// Watchlist Handlers

// GetWatchlists handles GET /watchlists
func GetWatchlists(c *fiber.Ctx) error {
	// TODO: Implement actual watchlist fetching
	return c.JSON(fiber.Map{
		"watchlists": []fiber.Map{
			{
				"id":          uuid.New().String(),
				"name":        "My DeFi Tokens",
				"description": "Favorite DeFi tokens to track",
				"items": []fiber.Map{
					{
						"type":       "token",
						"identifier": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
						"chainId":    1,
						"metadata": fiber.Map{
							"symbol": "WETH",
							"name":   "Wrapped Ether",
						},
					},
				},
				"createdAt": time.Now().Add(-30 * 24 * time.Hour),
				"updatedAt": time.Now().Add(-1 * 24 * time.Hour),
			},
		},
	})
}

// CreateWatchlist handles POST /watchlists
func CreateWatchlist(c *fiber.Ctx) error {
	// TODO: Implement actual watchlist creation
	var req fiber.Map
	if err := c.BodyParser(&req); err != nil {
		return errors.BadRequest("Invalid request body")
	}

	return c.Status(201).JSON(fiber.Map{
		"id":          uuid.New().String(),
		"name":        req["name"],
		"description": req["description"],
		"items":       req["items"],
		"createdAt":   time.Now(),
		"updatedAt":   time.Now(),
	})
}

// GetWatchlist handles GET /watchlists/:watchlistId
func GetWatchlist(c *fiber.Ctx) error {
	watchlistId := c.Params("watchlistId")
	// TODO: Implement actual watchlist fetching
	return c.JSON(fiber.Map{
		"id":          watchlistId,
		"name":        "My DeFi Tokens",
		"description": "Favorite DeFi tokens to track",
		"items":       []fiber.Map{},
		"createdAt":   time.Now().Add(-30 * 24 * time.Hour),
		"updatedAt":   time.Now(),
	})
}

// UpdateWatchlist handles PUT /watchlists/:watchlistId
func UpdateWatchlist(c *fiber.Ctx) error {
	watchlistId := c.Params("watchlistId")
	// TODO: Implement actual watchlist update
	var req fiber.Map
	if err := c.BodyParser(&req); err != nil {
		return errors.BadRequest("Invalid request body")
	}

	return c.JSON(fiber.Map{
		"id":          watchlistId,
		"name":        req["name"],
		"description": req["description"],
		"items":       req["items"],
		"updatedAt":   time.Now(),
	})
}

// DeleteWatchlist handles DELETE /watchlists/:watchlistId
func DeleteWatchlist(c *fiber.Ctx) error {
	// TODO: Implement actual watchlist deletion
	return c.SendStatus(204)
}