# spec/openapi.yaml
openapi: 3.1.0
info:
  title: DeFi Dashboard API
  version: 1.0.0
  description: Backend API for DeFi portfolio tracking and management
  contact:
    name: API Support
    email: api@defi-dashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.defi-dashboard.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    chainId:
      name: chainId
      in: query
      description: Blockchain network ID
      schema:
        type: integer
        enum: [1, 10, 56, 137, 8453, 42161]
    address:
      name: address
      in: path
      description: Ethereum address
      required: true
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - code
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Token:
      type: object
      properties:
        address:
          type: string
        symbol:
          type: string
        name:
          type: string
        decimals:
          type: integer
        chainId:
          type: integer
        logoUri:
          type: string
        price:
          type: number
        priceChange24h:
          type: number
      required:
        - address
        - symbol
        - decimals
        - chainId
...
    YieldPool:
      type: object
      properties:
        id:
          type: string
        protocol:
          type: string
        name:
          type: string
        chainId:
          type: integer
        apy:
          type: number
        tvl:
          type: number
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/Token'
        risks:
          type: array
          items:
            type: string
            enum: [impermanent_loss, smart_contract, oracle, liquidation]
      required:
        - id
        - protocol
        - name
        - chainId
        - apy
        - tvl
        - tokens
    YieldPosition:
      type: object
      properties:
        pool:
          $ref: '#/components/schemas/YieldPool'
        balance:
          type: string
        balanceUsd:
          type: number
        rewards:
          type: array
          items:
            type: object
            properties:
              token:
                $ref: '#/components/schemas/Token'
              amount:
                type: string
              amountUsd:
                type: number
        entryTime:
          type: string
          format: date-time
        pnl:
          type: number
        pnlPercentage:
          type: number
      required:
        - pool
        - balance
        - balanceUsd
    BridgeRoute:
      type: object
      properties:
        id:
          type: string
        fromChain:
          type: integer
        toChain:
          type: integer
        fromToken:
          $ref: '#/components/schemas/Token'
        toToken:
          $ref: '#/components/schemas/Token'
        bridge:
          type: string
        estimatedTime:
          type: integer
        fee:
          type: number
        feeUsd:
          type: number
      required:
        - id
        - fromChain
        - toChain
        - fromToken
        - toToken
        - bridge
        - estimatedTime
        - fee
    PnLExport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalPnl:
          type: number
        totalPnlPercentage:
          type: number
        transactions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              type:
                type: string
              asset:
                type: string
              amount:
                type: number
              price:
                type: number
              pnl:
                type: number
              fees:
                type: number
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [price_above, price_below, large_transfer, approval, liquidity_change]
        status:
          type: string
          enum: [active, triggered, expired, disabled]
        target:
          type: object
          properties:
            type:
              type: string
              enum: [token, address, pool]
            identifier:
              type: string
            chainId:
              type: integer
        conditions:
          type: object
        notification:
          type: object
          properties:
            email:
              type: boolean
            webhook:
              type: string
        createdAt:
          type: string
          format: date-time
        triggeredAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - status
        - target
        - conditions
        - createdAt
    Watchlist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [token, address, pool]
              identifier:
                type: string
              chainId:
                type: integer
              metadata:
                type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - items
        - createdAt
security:
  - bearerAuth: []
paths:
...
  /yield/pools:
    get:
      tags:
        - yield
      summary: Get available yield pools
      operationId: getYieldPools
      parameters:
        - $ref: '#/components/parameters/chainId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: protocol
          in: query
          schema:
            type: string
        - name: minApy
          in: query
          schema:
            type: number
        - name: minTvl
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [apy, tvl, name]
            default: apy
      responses:
        '200':
          description: Yield pools
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/YieldPool'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /yield/positions/{address}:
    get:
      tags:
        - yield
      summary: Get yield positions
      operationId: getYieldPositions
      parameters:
        - $ref: '#/components/parameters/address'
        - $ref: '#/components/parameters/chainId'
      responses:
        '200':
          description: Yield positions
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalValue:
                    type: number
                  totalPnl:
                    type: number
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/YieldPosition'
  /bridge/routes:
    post:
      tags:
        - bridge
      summary: Get bridge routes
      operationId: getBridgeRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromChain:
                  type: integer
                toChain:
                  type: integer
                fromToken:
                  type: string
                toToken:
                  type: string
                amount:
                  type: string
              required:
                - fromChain
                - toChain
                - fromToken
                - amount
      responses:
        '200':
          description: Available bridge routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/BridgeRoute'
  /swap/quote:
    post:
      tags:
        - swap
      summary: Get swap quotes
      operationId: getSwapQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapQuoteRequest'
      responses:
        '200':
          description: Swap quotes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SwapRoute'
  /swap/execute:
    post:
      tags:
        - swap
      summary: Execute swap transaction
      operationId: executeSwap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                routeId:
                  type: string
                userAddress:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
              required:
                - routeId
                - userAddress
      responses:
        '200':
          description: Swap executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  txHash:
                    type: string
                required:
                  - txHash
  /analytics/export/{address}:
    get:
      tags:
        - analytics
      summary: Export P&L report
      operationId: exportPnL
      parameters:
        - $ref: '#/components/parameters/address'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: P&L report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PnLExport'
            text/csv:
              schema:
                type: string
  /alerts:
    get:
      tags:
        - alerts
      summary: Get user alerts
      operationId: getAlerts
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, triggered, expired, disabled]
      responses:
        '200':
          description: User alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - alerts
      summary: Create alert
      operationId: createAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [price_above, price_below, large_transfer, approval, liquidity_change]
                target:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [token, address, pool]
                    identifier:
                      type: string
                    chainId:
                      type: integer
                conditions:
                  type: object
                notification:
                  type: object
                  properties:
                    email:
                      type: boolean
                    webhook:
                      type: string
              required:
                - type
                - target
                - conditions
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          $ref: '#/components/responses/BadRequest'
  /alerts/{alertId}:
    get:
      tags:
        - alerts
      summary: Get alert by ID
      operationId: getAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - alerts
      summary: Update alert
      operationId: updateAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, disabled]
                conditions:
                  type: object
                notification:
                  type: object
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - alerts
      summary: Delete alert
      operationId: deleteAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Alert deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /watchlists:
    get:
      tags:
        - watchlist
      summary: Get user watchlists
      operationId: getWatchlists
      responses:
        '200':
          description: User watchlists
          content:
            application/json:
              schema:
                type: object
                properties:
                  watchlists:
                    type: array
                    items:
                      $ref: '#/components/schemas/Watchlist'

    post:
      tags:
        - watchlist
      summary: Create watchlist
      operationId: createWatchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [token, address, pool]
                      identifier:
                        type: string
                      chainId:
                        type: integer
                      metadata:
                        type: object
              required:
                - name
      responses:
        '201':
          description: Watchlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '400':
          $ref: '#/components/responses/BadRequest'
  /watchlists/{watchlistId}:
    get:
      tags:
        - watchlist
      summary: Get watchlist by ID
      operationId: getWatchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Watchlist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - watchlist
      summary: Update watchlist
      operationId: updateWatchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                items:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Watchlist updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - watchlist
      summary: Delete watchlist
      operationId: deleteWatchlist
      parameters:
        - name: watchlistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Watchlist deleted
        '404':
          $ref: '#/components/responses/NotFound'
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "VALIDATION_ERROR"
            message: "Invalid request parameters"
            details:
              field: "address"
              issue: "Invalid Ethereum address format"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
...